traffix-bot/pyproject.toml
[project]
name = "traffix-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
   "dotenv>=0.9.9",
   "fastapi>=0.121.1",
   "google-genai>=1.50.0",
   "groq>=0.34.0",
   "pinecone>=7.3.0",
   "pymupdf>=1.26.6",
   "uvicorn>=0.38.0",
]
